requirements.txt
adal==1.2.7
altgraph==0.17.3
azure-core==1.26.4
azure-identity==1.12.0
azure-kusto-data==4.1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
dpath==2.1.5
gitignore-parser==0.1.3
idna==3.4
ijson==3.2.0.post0
luau==0.2.0
lxml==4.9.2
midas-data-util==0.2.3
msal==1.21.0
msal-extensions==1.0.0
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.2
pandas==2.0.0
pbit==0.2.6
pefile==2023.2.7
portalocker==2.7.0
pyaml==21.10.1
pycparser==2.21
pyinstaller==5.10.0
pyinstaller-hooks-contrib==2023.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
types-toml==0.10.8.6
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
